Switch to use Action Provider Tools Blueprint
'add_action_routes_to_blueprint' is being removed in an upcoming
release. The activities which it performed are better encapsulated by
'ActionProviderBlueprint', which is available in all recent versions
of Action Provider Tools.

The blueprint provides registration methods which can be used as
decorators, but to preserve the semantics of the current application
as much as possible, I have refrained from using them as such. The
registration calls are made imperatively in the app factory, similar
to how the `add_action_routes_to_blueprint` function behaved.

The ActionProviderBlueprint looks up client credentials from app
config based on the name of the blueprint, so the client credential
names are set appropriately in app config to match.